Define AeroelasticModel abstract base class
This commit defines an AeroelasticModel abstract base class in order to
adhere to the Open-Closed Principle (OCP) and Dependency Inversion
Principle (DIP).

In doing so, useful attributes and methods are also defined for all
specialized aeroelastic models. These include a state-space formulation
and simulation. In addition, it includes a solver based approach to
calculating the flutter and divergence speeds.