Add optimization framework and restructure code
This commit adds a simple optimization framework to simplify the setup
of optimization problems. Also, it moves the optimization definitions
to the end of the file such that first the mathematical models are
created, followed by the optimization definitions. This is done to
group the code by function rather than by assignment task.